advocacia_saas/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.0.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.0.0
Pillow>=10.0.0
gunicorn==21.2.0
psycopg2-binary>=2.9.0
reportlab==4.4.7
mercadopago==2.3.0
xhtml2pdf==0.2.17
bleach==6.1.0
openai>=1.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-Caching==2.3.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
sentry-sdk[flask]==1.40.0
redis==5.0.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
boto3==1.34.0
Flask-SocketIO==5.5.1
python-socketio==5.15.1
markdown==3.5.2

Flask-Mail==0.9.1

python-dateutil==2.8.2
aiohttp==3.13.2
